+++ qbft/timeout.go
@@ -10,9 +10,9 @@ import (
 var (
 	quickTimeoutThreshold = Round(8) //nolint
 	// quickTimeout is the timeout in seconds for the first 8 rounds
-	quickTimeout uint64 = 2 // 2 seconds
+	quickTimeout int64 = 2 // 2 seconds
 	// slowTimeout is the timeout in seconds for rounds after the first 8
-	slowTimeout uint64 = 120 // 2 minutes
+	slowTimeout int64 = 120 // 2 minutes
 	// CutoffRound which round the instance should stop its timer and progress no further
 	CutoffRound = 15 // stop processing instances after 8*2+120*6 = 14.2 min (~ 2 epochs)
 )
@@ -25,15 +25,15 @@ type RoundTimer struct {
 	// network is the beacon network
 	Network types.BeaconNetwork
 	//current unix epoch time in seconds
-	CurrentTime uint64
+	CurrentTime int64
 }
 
-func (t *RoundTimer) TimeoutForRound(round Round) uint64 {
+func (t *RoundTimer) TimeoutForRound(round Round) int64 {
 	switch t.Role {
 	case types.BNRoleAttester, types.BNRoleSyncCommittee:
-		return AttestationOrSyncCommitteeTimeout(round, t.Height, t.Network) - t.CurrentTime
+		return max(AttestationOrSyncCommitteeTimeout(round, t.Height, t.Network)-t.CurrentTime, 0)
 	case types.BNRoleAggregator, types.BNRoleSyncCommitteeContribution:
-		return AggregationOrContributionTimeout(round, t.Height, t.Network) - t.CurrentTime
+		return max(AggregationOrContributionTimeout(round, t.Height, t.Network)-t.CurrentTime, 0)
 	default:
 		return DefaultTimeout(round)
 	}
@@ -41,12 +41,12 @@ func (t *RoundTimer) TimeoutForRound(round Round) uint64 {
 
 // RoundTimeout returns the unix epoch time (seconds) in which we should send a RC message
 // Called for all beacon duties other than proposals
-func RoundTimeout(round Round, height Height, baseDuration uint64, network types.BeaconNetwork) uint64 {
+func RoundTimeout(round Round, height Height, baseDuration int64, network types.BeaconNetwork) int64 {
 	// Calculate additional timeout in seconds based on round
-	var additionalTimeout uint64
-	additionalTimeout = uint64(min(round, quickTimeoutThreshold)) * quickTimeout
+	var additionalTimeout int64
+	additionalTimeout = int64(min(round, quickTimeoutThreshold)) * quickTimeout
 	if round > quickTimeoutThreshold {
-		slowPortion := uint64(round-quickTimeoutThreshold) * slowTimeout
+		slowPortion := int64(round-quickTimeoutThreshold) * slowTimeout
 		additionalTimeout += slowPortion
 	}
 
@@ -54,24 +54,24 @@ func RoundTimeout(round Round, height Height, baseDuration uint64, network types
 	timeoutDuration := baseDuration + additionalTimeout
 
 	// Get the unix epoch start time of the duty seconds
-	dutyStartTime := uint64(network.EstimatedTimeAtSlot(phase0.Slot(height)))
+	dutyStartTime := int64(network.EstimatedTimeAtSlot(phase0.Slot(height)))
 
 	// Calculate the time until the duty should start plus the timeout duration
 	return dutyStartTime + timeoutDuration
 }
 
 // AttestationOrSyncCommitteeTimeout returns the unix epoch time (seconds) in which we should send a RC message
-func AttestationOrSyncCommitteeTimeout(round Round, height Height, network types.BeaconNetwork) uint64 {
+func AttestationOrSyncCommitteeTimeout(round Round, height Height, network types.BeaconNetwork) int64 {
 	return RoundTimeout(round, height, 4, network)
 }
 
 // AggregationOrContributionTimeout returns the unix epoch time (seconds) in which we should send a RC message
-func AggregationOrContributionTimeout(r Round, height Height, network types.BeaconNetwork) uint64 {
+func AggregationOrContributionTimeout(r Round, height Height, network types.BeaconNetwork) int64 {
 	return RoundTimeout(r, height, 8, network)
 }
 
 // DefaultTimeout returns the duration in seconds in which we should send a RC message
-func DefaultTimeout(round Round) uint64 {
+func DefaultTimeout(round Round) int64 {
 	if round <= quickTimeoutThreshold {
 		return quickTimeout
 	} else {
@@ -106,6 +106,14 @@ func (i *Instance) UponRoundTimeout() error {
 // min returns the minimum of two values
 func min[T constraints.Ordered](a, b T) T {
 	if a < b {
+		return a
+	}
+	return b
+}
+
+// max returns the maximum of two values
+func max[T constraints.Ordered](a, b T) T {
+	if a > b {
 		return a
 	}
 	return b
